Rename 3D reconstruction experiment and undo changes to afarbin's files.
